QmlDebug: Support sending multiple messages per packet
This reduces the overhead caused by the prepending of service names to
packets. When running the planets example from qtcanvas3d through
qmlprofiler the average message length, without service name, is 26
bytes. The average full packet, with name and length, is 64 bytes long.
In fact, the time between stopping the application and the last message
arriving in the profiling client is reduced by 30-50% with this
change.

Change-Id: I0ffdd0483d45bbe8b092c59a9dcd63c6dc59119c
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>