Particles: factor out t^2.
According to IEEE754 a * t * t is not the same as a * (t * t), hence the
compiler won't lift t^2 out as common expression when in another place
'a' is replaced with 'b'. Now for the particles this doesn't matter as
it doesn't need super high precision, nor should any of the calculations
come in overflow territory. So by manually introducing t_sq(uare), the
compiler can remove duplicate t^2 calculations after inlining.

Change-Id: Ibfdb054e34945d7e78eb993fdcedfae886472e83
Reviewed-by: Lars Knoll <lars.knoll@theqtcompany.com>