QQuickItemView::forceLayout: Also call layout when d->forceLayout
This way callers of forceLayout get a relayout if the dimension
of a delegate that affects the itemview content size has just changed
but the itemview content size has not been updated yet.

This is useful because sometimes in the QML side you want to do things when the
item view delegate size has changed but need the item view content size to be already updated.

Change-Id: I846984a841e8e14c84d7a700a7ff736196b60afb
Reviewed-by: Michael Brasser <michael.brasser@live.com>
Reviewed-by: Shawn Rutledge <shawn.rutledge@theqtcompany.com>