QQuickMenu: replace eventFilter() with keyReleaseEvent()
We are now creating an internal QQuickPopupItem that always exists, and we
forward its events to QQuickPopup to make it behave like it was an item
itself. Before, we had to filter a user/style supplied content item, but now
it's possible to override the event handlers of the popup item.

Change-Id: I909f82c4e09f7de92d30981d6366d02590a708f3
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>