QQuickMenu: eliminate the internal dummyFocusItem
Thanks to the event handlers and itemChange() that were introduced in the
QQuickPopup base class, QQuickMenu is now able to control the focus so that
a dummy focus item is no longer necessary.

The purpose of the dummy focus item was to hold active focus when the menu
becomes visible, so that none of the actual menu items have active focus until
the user starts navigating with keys. Instead of using such dummy focus item,
we can now clear "sub" focus in menu's focus scope.

Change-Id: I664ef2d314244f55a8e1e8e77094dcc43e59953f
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>