D3D12: Fix async texture upload logic
The cleaning of the staging buffer and others happened too soon in the
original version of the code. We can only be sure the upload has finished
when the frame waiting (on the GPU) for the relevant uploads to finish on
the copy queue has finished on the 3D queue. That we only know in frame + 2's
beginFrame().

Change-Id: Ie77e4664cc2f5308a5f7b9ac8dc438ca78e70b1b
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>