tst_material/universal: use the appropriate style
The main purpose of these tests were to test the attached property
and its inheritance, so it didn't matter which particular style was
active while running the tests. However, it simply looks better if
they run with the appropriate style, and this also allows us to add
style-specific tests (such as font attributes).

Change-Id: I827f26e469bce3445ffe7e24d6299bcfb587d849
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>
Reviewed-by: Liang Qi <liang.qi@theqtcompany.com>