Change style inheritance for popups
The decision to make popups inherit theme & accent from its parent item
was driven by ComboBox. However, in many cases it has an undesired side
effect, so we've changed Popup to inherit its style attributes from the
parent window instead, just like we did for fonts. The only exception
to this is ComboBox, where the popup is an integral part of the control.
This special case is now handled in the respective style implementation.

A concrete example is that we can now specify dark theme by default for
Material style ToolBar to get a better matching light text against the
colorful background. In Gallery, this won't effect the options menu,
which is a child of a ToolButton. The menu retains light theme along the
rest of the application.

Change-Id: Ibdc8fcf5b5fa258d853410a9b40368472424a8c6
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>