QQuickOverlay: fix background visibility for multiple modal popups
The idea of the "modalPopups > 1" check was to avoid calling QQmlProperty::write(),
but it missed the fact that one modal popup might be still closing while a new modal
popup is already opened. Thus, there's temporarily 2 modal popups visible, and we
must transition the overlay background according to the state of the last shown modal
popup.

Change-Id: If25d83ccaa5adc29a9bc8d660f41fb66f90fb167
Task-number: QTBUG-51916
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>