D3D12: Support styled and outlined text
The texture resizing highlighted a potential issue with deferred deletes
and texture releases: All the texture management functions are special in
the sense that they can be called outside a begin-endFrame. (this is what
glyph caches do in fact when resizing is involved) In this case using
currentPFrameIndex (referring to the last frame's pframeData slot) seems safe,
but using it is in fact incorrect under some scenarios: a resource queued for
releasing between frame 1 and 2 cannot be released in beginFrame of frame 3,
which was the case until now. The CPU wait in frame 3's beginFrame only guarantees
the commands submitted before frame 1's endFrame have completed, and those will not
include a wait for the copy queue's commands that are added between frames 1 and 2.

Using the next frame's pframeData is not an option either since that would
lead to an immediate release once beginFrame is called the next time.

Therefore, introduce an out-of-frame queue for deleting and texture slot reuse.
With the previous example the release will be done in frame 4, which is
safe since everything submitted in frame 2 - which will normally include a
wait for the copy queue if there were commands issued there before frame 2
using the to-be-released texture - are guaranteed to be done due to the
CPU wait.

Change-Id: Iaa3b5f0af991946dc97ba44448823b22cbf419dc
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>