Change the default value of QQuickPopup::closePolicy
When Popup::ClosePolicy was introduced, I thought that it would be
better to not set OnPressOutside by default, because not all popups
necessarily want to close on press outside. However, I've realized
that every time I do

    Button {
        onClicked: popup.open()
        Popup { }
    }

...there's literally no interactive way to close the popup (popups
don't take focus by default, so Esc doesn't work).

Provided that a top-level QWidget with the Qt::Popup flag set also
closes automatically on press outside, it probably makes more sense
for us to follow the same pattern so that poor new users don't have
to lookup the docs just to figure out how to make a popup close. :)

In this case it's less troublesome to specify a different default
value in those special scenarios/subclasses where closing on press
outside is not desired (eg. the upcoming Popup-based Drawer).

Change-Id: Iaeab4c548d63728df12232d8b65d469f0a2d80f5
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>