D3D12: Fix incorrect buffer update logic
Not updating totalDirtyInFrame when the range is already in dirty
is wrong because the range may not have been added to totalDirtyInFrame
yet. The issue becomes visible when animated items stop animating after
some time due to not syncing constant buffer updates from the previous,
in-flight frame in every other frame. The result was heavy stutter due to
jumping back and forth between the last and the last - 1 animation
step's matrix.

While we are at it, beginFrame's dirty synchronization is made more robust
and more correct for the (theoretical) case MAX_FRAMES_IN_FLIGHT > 2.

Change-Id: Ifeb900b2e1bd206f6e8f3c370d45b3d831d602f0
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>