Add Control::activeKeyFocus
This is a convenience property that combines Item::activeFocus and
Control::focusReason. This allows us to solve QTBUG-51796 in an elegant
way - we can simply set Qt::StrongFocus policy on most of the interactive
controls, such as buttons. The only thing we need to make sure is to use
activeKeyFocus instead of activeFocus in the styles. This ensures that key
focus is only visualized when actually interacting with keys - not when
interacting via touch or mouse. This is a generalized version of the
solution that was already used in the Universal style ApplicationWindow.

Change-Id: Ifbf78e3e412f3791c8f7c369bb2de53af9ac6b0f
Task-number: QTBUG-51796
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>