D3D12: Say hello to the threaded render loop
Adapt the threaded OpenGL render loop. The inner workings are pretty much
the same. There are renamings, cleanups and simplifications here and there
since life is somewhat easier without having to juggle contexts and the insanity
of fallback surfaces.

The engine is now changed to be independent of QWindow as the per-render-thread
engine instance lives on the per-window render thread from now on.

There may be some glitches here and there, but the casess from the nodetypes
test are functional already.

Device loss handling is improved (the scenegraph is now attempted to be
reset) but is likely unstable for now.

Change-Id: Icaf432740f4ffa8932a444abc30c8cbce395ffcd
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>