D3D12: Fix disappearing text on resize
Geometry rebuilds resulted in the disappearance of some text items. This
is due to the cbuffer zeroing that happened on DirtyAll without writing
out all the data again afterwards. In fact the entire memsetting is
superfluous and useless in practice - all the m_lastXxxx branches are hit
due to DirtyOpacity anyhow. So just remove the ill-conceived
DirtyConstantBuffer flag.

Change-Id: Icdc534e144031eb47ee2bfc08a16c502d9f1d9b0
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>