D3D12: Fix text material type
We have a single material class for all styles and glyph formats. type()
takes the style correctly into account, but formats that need blending
also need special handling since such a material is different from another
material with the same style but different glyph format from the
renderer's point of view. In practice this won't be hit since the glyphs'
formats are unlikely to be different, but let's do it correctly nonetheless.

Change-Id: If7250678a6bbf486a44ba05e351f7622401f7106
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>