QML: do not re-use the resolver data for members.
On successful lookup, the resolver data was cleared and re-used for the
resolved member.

The effect is that a second time the resolver is used, it will not be
able to do the same lookup, resulting in it returning an unknown type.
Because the same expression might need to be resolved multiple times
(e.g. when a dependency changes type), this results in the wrong type
(var), and then more iterations to propagate this wrong type to all
usages.

Instead, return a new resolver with its own data for the resolved
member. This way, a member access on this result can be resolved
correctly by this new resolver.

Change-Id: Ia930c08a2e4a2182d800192547fc03cba209c78c
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>