D3D12: Fix scissor rect
It was off by one due to the historical behavior of QRect::right() and
bottom(). Originally the issue was worked around in the renderer. It is
now done properly in queueScissor instead.

In addition, draw calls with a zero scissor result in a debug layer
warning. It is not necessary to add the draw call related commands to the
command list in this case so just skip them.

Change-Id: I016cfbecc093598aa0825bbc7811934b05882486
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>