D3D12: Fix another set of disappearing text cases
Even after the previous fixes the cbuffer did not get updated after a
rebuild in some cases. (Re)introduce a dirty state - this time with the
proper name - indicating that any cached material data should be treated
as invalid. This allows updatePipeline to disregard cached state and
update the cbuffer even for values that are not otherwise dependent on
the opacity or the matrix.

Change-Id: I8e4e6c52067d1db4bd7049960bc8cfc7cac0b0ef
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>