Add imageSource property to attached Drag property
This property holds a URL which is loaded into a QPixmap via
QQuickPixmap, and subsequently set on the QDrag object.

This is especially important for drag and drop to external applications,
where of course a QQuickItem cannot be dragged; however, combined with
QQuickItem::grabToImage(), this change allows setting a QQuickItem's
contents as the drag pixmap.

[ChangeLog][QtQuick][Drag] Added imageSource property to attached Drag
object: this allows drag sources to specify the pixmap to be drawn next
to the mouse cursor when starting a drag and drop operation.

Task-number: QTBUG-37366
Change-Id: Ibcf1d888c525e50976a116ec743ce6fca4c31525
Reviewed-by: Michael Brasser <michael.brasser@live.com>