Add QQuickAbstractButton::down
This property will determine whether or not the button is visually
pressed. Having such a distinction allows users more control over their
controls.

The patch also fixes the problem with ComboBox where pressing on the
ComboBox when it's open would cause a delegate in the popup to show as
being pressed.

Unless explicitly set, this property follows the value of the pressed
property. To return to the default value, set it to undefined.

Change-Id: I29ecf325ed2ede125613f0c878b0427937599866
Task-number: QTBUG-51005
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>