Material: add foreground and background attached properties
This has been a common request on the mailing lists and forums. Users
want to customize the text color of various controls without having to
replace whole delegates. This change makes it possible to specify both
foreground (eg. Button text) and background (eg. ApplicationWindow,
Page and Pane background) colors so that they propagate to children in
the same way than accent, primary and theme do.

Change-Id: Idae52650ed8620beeff7bf4c8263ffdc1f6ef728
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>