2DRenderer: Mark all regions of deleted node as dirty
Previous behavior was to always subtract the current bounding rect of
items when calculating the previous dirty region.  When you remove a
node though, there is no current bounding rect because the item no
longer exists.  The last valid boundingRect was use instead, which led
to the previous bounding rect to not get marked as dirty when the node
was removed.

Change-Id: Ic1a4f872d4d46125b06e2588aae695b8ff67d0fd
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>