Dial: add wrap property
This works by preventing changes in position that are considered
too large. A change is considered too large when the difference
between the old and new positions is greater than or equal to 0.5
*and* the position of the mouse is below the center of the dial.

This effectively makes it impossible to drag the dial from left
to right when the mouse is below its center, but still allows
doing so above its center.

This is useful for applications such as volume dials, where a
large change can be dangerous.

Change-Id: I1d7800e5ba16dbe0642974b8e53d8fcd921a01d7
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>