Universal: add foreground and background attached properties
Same as b51e520f for Material.

This has been a common request on the mailing lists and forums. Users
want to customize the text color of various controls without having to
replace whole delegates. This change makes it possible to specify both
foreground (eg. Button text) and background (eg. ApplicationWindow,
Page and Pane background) colors so that they propagate to children in
the same way than accent and theme do.

Change-Id: I55084c58f0c62e8088b270ed98437241e382ca15
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>