QQuickTextInput: Listen for changes to cursorFlashTime
cursorFlashTime will now change dynamically from QPA while platform
controlled text selection (on mobile) is ongoing. This patch
will therefore update QWidgetLineControl so that it listens to the
cursorFlashTimeChanged signal and changes the blinking rate when
triggered.

The previous code had a function setCursorBlinkPeriod, which is now
changed to setBlinkingCursorEnabled (like in QQuickTextControlPrivate).
This is because all callers of the function did either pass
"QApplication::cursorFlashTime" or "0", which basically means enable
or disable blinking. This moves the control of the blinking rate
fully to updateCursorFlashTime, which simplifies the code a
bit, especially when cursorFlashTime can change.

Note that when setting a blink period to 0, it means "show the
cursor without blinking". AFAICS, the current implementation did
not guarantee that. This is now made more explicit in the code. If
hiding the cursor is needed, a separate function "setCursorVisible"
is already available for controlling that.

Change-Id: Ie3222525474e03b485ff8585fd8d7da6fd5b26e5
Reviewed-by: Jan Arve Sæther <jan-arve.saether@theqtcompany.com>