Adaptable shader effect enablers
The D3D12 node implementation is mostly missing. The rest of the enablers
should be in place now however.

Importing QtQuick 2.8 provides new properties for ShaderEffect:

- shaderType can be used to decide which language should be used

- shaderCompilationType tells if compilation is runtime or offline

- shaderSourceType tells if the vertex/fragmentShader properties refer
  to source strings or source files or pre-compiled bytecode files

The last two are bitmasks. In practice however we will support only
one approach per backend for now (runtime + source string for OpenGL,
offline + bytecode for D3D12).

In addition to QSGShaderEffectNode, introduce the QSGGuiThreadShaderEffectManager
interface. This provides the gui thread bits for the above and performs shader
reflection. Backends that use the new ShaderEffect system must provide both.
For each ShaderEffect item there will be a manager (on the gui thread) and
a node (on the render thread).

Reflection is expected to be done via standard helper libs (d3d12shader
+ D3DReflect from d3dcompiler for the D3D12 backend), or via manually inspecting
SPIR-V, or parsing the source (like the GL path does now), or by using some
3rd party library (not recommended). In any case we require that reflection is
doable on the gui thread without dependencies to the actual graphics API.

The ShaderEffect documentation is greatly extended, covering HLSL and the new
properties.

The test app uses manually compiled shaders on its HLSL path for now. This
is because there is no story yet for build system integration for public
use (the internal HLSL -> bytecode in C header rule is only suitable for
the d3d12 plugin itself, apps need something different).

Change-Id: Id112104906fbcb26b9902a35f19d8d509b340d1b
Reviewed-by: Andy Nichols <andy.nichols@qt.io>